+++ src/main/java/net/glowstone/net/message/play/entity/SpawnPlayerMessage.java
@@ -12,7 +12,7 @@ public final class SpawnPlayerMessage implements Message {
 
     private final int id;
     private final UUID uuid;
-    private final int x, y, z;
+    private final double x, y, z;
     private final int rotation, pitch;
     private final List<MetadataMap.Entry> metadata;
